mgr/dashboard_v2: Add RBD mirroring functionality
This commit adds the RBD mirroring functionality of the first dashboard to
the dashboard_v2 REST API. That's why the code is basically based on the
code of the first dashboard. It was just brought into a different structure
to fit into the structure of dashboard_v2.

The new API endpoint can be found at http://<host>:<port>/api/rbdmirror

The commit adds also related tests.

Signed-off-by: Tatjana Dehler <tdehler@suse.com>